+++ src/app/containers/AccountContainer.js
@@ -1,14 +1,50 @@
 import React from 'react';
-import { UsernameSection, NewsSection, DeleteSection, EmailSection, ObserverSections } from 'react-components';
+import { UsernameSection, NewsSection, DeleteSection, EmailSection } from 'react-components';
+import { c } from 'ttag';
+
+import Page from '../components/Page';
+
+export const getAccountPage = () => {
+    return {
+        text: c('Title').t`Account`,
+        route: '/settings/account',
+        sections: [
+            {
+                text: c('Title').t`Username`,
+                id: 'username'
+            },
+            {
+                text: c('Title').t`Passwords`,
+                id: 'passwords'
+            },
+            {
+                text: c('Title').t`Two factor authentication`,
+                id: '2fa'
+            },
+            {
+                text: c('Title').t`Recovery & notification`,
+                id: 'email'
+            },
+            {
+                text: c('Title').t`Email subscriptions`,
+                id: 'news'
+            },
+            {
+                text: c('Title').t`Delete account`,
+                id: 'delete'
+            }
+        ]
+    };
+};
 
 const AccountContainer = () => {
     return (
-        <ObserverSections>
-            <UsernameSection id="username" />
-            <EmailSection id="email" />
-            <NewsSection id="news" />
-            <DeleteSection id="delete" />
-        </ObserverSections>
+        <Page config={getAccountPage()}>
+            <UsernameSection />
+            <EmailSection />
+            <NewsSection />
+            <DeleteSection />
+        </Page>
     );
 };
