+++ src/app/containers/AppearanceContainer.js
@@ -1,13 +1,37 @@
 import React from 'react';
-import { LayoutsSection, ThemesSection, ObserverSections, ToolbarsSection } from 'react-components';
+import { LayoutsSection, ThemesSection, ToolbarsSection } from 'react-components';
+import { c } from 'ttag';
+
+import Page from '../components/Page';
+
+export const getAppearancePage = () => {
+    return {
+        text: c('Title').t`Appearance`,
+        route: '/settings/appearance',
+        sections: [
+            {
+                text: c('Title').t`Layouts`,
+                id: 'layouts'
+            },
+            {
+                text: c('Title').t`Toolbars`,
+                id: 'toolbars'
+            },
+            {
+                text: c('Title').t`Themes`,
+                id: 'themes'
+            }
+        ]
+    };
+};
 
 const AppearanceContainer = () => {
     return (
-        <ObserverSections>
-            <LayoutsSection id="layouts" />
-            <ToolbarsSection id="toolbars" />
-            <ThemesSection id="themes" />
-        </ObserverSections>
+        <Page config={getAppearancePage()}>
+            <LayoutsSection />
+            <ToolbarsSection />
+            <ThemesSection />
+        </Page>
     );
 };
