+++ src/app/containers/AppsContainer.js
@@ -1,4 +1,6 @@
 import React from 'react';
+import { c } from 'ttag';
+import { PERMISSIONS } from 'proton-shared/lib/constants';
 import {
     ProtonMailBridgeSection,
     ProtonVPNAppsSection,
@@ -7,15 +9,44 @@ import {
 } from 'react-components';
 
 import Page from '../components/Page';
-import { getAppsPage } from '../pages/apps';
 
-const COMPONENTS = {
-    'protonmail-bridge': ProtonMailBridgeSection,
-    'protonmail-apps': ProtonVPNAppsSection,
-    'protonmail-beta': ProtonMailBetaSection,
-    'protonvpn-apps': ProtonMailAppsSection
+const { PAID_MAIL } = PERMISSIONS;
+
+export const getAppsPage = () => {
+    return {
+        text: c('Title').t`Apps`,
+        route: '/settings/apps',
+        sections: [
+            {
+                text: c('Title').t`ProtonMail Bridge`,
+                id: 'protonmail-bridge',
+                permissions: [PAID_MAIL]
+            },
+            {
+                text: c('Title').t`ProtonMail Apps`,
+                id: 'protonmail-apps'
+            },
+            {
+                text: c('Title').t`ProtonMail Beta`,
+                id: 'protonmail-beta'
+            },
+            {
+                text: c('Title').t`ProtonVPN Apps`,
+                id: 'protonvpn-apps'
+            }
+        ]
+    };
 };
 
-const AppsContainer = () => <Page page={getAppsPage()} components={COMPONENTS} />;
+const AppsContainer = () => {
+    return (
+        <Page config={getAppsPage()}>
+            <ProtonMailBridgeSection />
+            <ProtonVPNAppsSection />
+            <ProtonMailBetaSection />
+            <ProtonMailAppsSection />
+        </Page>
+    );
+};
 
 export default AppsContainer;
