+++ src/app/containers/AutoReplyContainer.js
@@ -1,10 +1,23 @@
 import React from 'react';
-import { ObserverSections, AutoReplySection, RelatedSettingsSection } from 'react-components';
+import { AutoReplySection, RelatedSettingsSection } from 'react-components';
 import { c } from 'ttag';
+import { PERMISSIONS } from 'proton-shared/lib/constants';
+
+import Page from '../components/Page';
+
+const { PAID_MAIL } = PERMISSIONS;
+
+export const getAutoReply = () => {
+    return {
+        text: c('Title').t`Auto-reply`,
+        route: '/settings/auto-reply',
+        permissions: [PAID_MAIL]
+    };
+};
 
 const AutoReplyContainer = () => {
     return (
-        <ObserverSections>
+        <Page config={getAutoReply()}>
             <AutoReplySection id="auto-reply" />
             <RelatedSettingsSection
                 id="related-settings"
@@ -18,7 +31,7 @@ const AutoReplyContainer = () => {
                     }
                 ]}
             />
-        </ObserverSections>
+        </Page>
     );
 };
