+++ src/app/containers/DomainsContainer.js
@@ -1,6 +1,29 @@
 import React from 'react';
 import { c } from 'ttag';
-import { DomainsSection, RelatedSettingsSection, CatchAllSection, ObserverSections } from 'react-components';
+import { DomainsSection, RelatedSettingsSection, CatchAllSection } from 'react-components';
+import { PERMISSIONS } from 'proton-shared/lib/constants';
+
+import Page from '../components/Page';
+
+const { ADMIN, PAID_MAIL } = PERMISSIONS;
+
+export const getDomainsPage = () => {
+    return {
+        text: c('Title').t`Custom domain`,
+        route: '/settings/domains',
+        permissions: [ADMIN, PAID_MAIL],
+        sections: [
+            {
+                text: c('Title').t`Custom domain`,
+                id: 'domains'
+            },
+            {
+                text: c('Title').t`Catch-all`,
+                id: 'catch-all'
+            }
+        ]
+    };
+};
 
 const DomainsContainer = () => {
     const relatedSettingsList = [
@@ -19,11 +42,11 @@ const DomainsContainer = () => {
         }
     ];
     return (
-        <ObserverSections>
+        <Page config={getDomainsPage()}>
             <DomainsSection id="domains" />
             <CatchAllSection id="catch-all" />
             <RelatedSettingsSection id="related-settings" list={relatedSettingsList} />
-        </ObserverSections>
+        </Page>
     );
 };
