+++ src/app/containers/FiltersContainer.js
@@ -1,7 +1,27 @@
 import React from 'react';
+import { c } from 'ttag';
+
+import Page from '../components/Page';
+
+export const getFiltersPage = () => {
+    return {
+        text: c('Title').t`Filters`,
+        route: '/settings/filters',
+        sections: [
+            {
+                text: c('Title').t`Custom filters`,
+                id: 'custom'
+            },
+            {
+                text: c('Title').t`Spam filters`,
+                id: 'spam'
+            }
+        ]
+    };
+};
 
 const FiltersContainer = () => {
-    return <h1>FiltersSection</h1>;
+    return <Page config={getFiltersPage()} />;
 };
 
 export default FiltersContainer;
