+++ src/app/containers/FoldersLabelsContainer.js
@@ -1,7 +1,17 @@
 import React from 'react';
+import { c } from 'ttag';
+
+import Page from '../components/Page';
+
+export const getLabelsPage = () => {
+    return {
+        text: c('Title').t`Folders/labels`,
+        route: '/settings/labels'
+    };
+};
 
 const FoldersLabelsContainer = () => {
-    return <h1>FoldersLabelsSection</h1>;
+    return <Page config={getLabelsPage()} />;
 };
 
 export default FoldersLabelsContainer;
