+++ src/app/containers/KeysContainer.js
@@ -1,7 +1,25 @@
 import React from 'react';
+import { c } from 'ttag';
 
-const KeysContainer = () => {
-    return <h1>KeysSection</h1>;
+import Page from '../components/Page';
+
+export const getKeysPage = () => {
+    return {
+        text: c('Title').t`Keys`,
+        route: '/settings/keys',
+        sections: [
+            {
+                text: c('Title').t`Email encryption keys`,
+                id: 'addresses'
+            },
+            {
+                text: c('Title').t`Contact encryption keys`,
+                id: 'user'
+            }
+        ]
+    };
 };
+
+const KeysContainer = () => <Page config={getKeysPage()}>KeysSection</Page>;
 
 export default KeysContainer;
