+++ src/app/containers/MembersContainer.js
@@ -1,8 +1,26 @@
 import React from 'react';
 import { MembersSection } from 'react-components';
+import { c } from 'ttag';
+import { PERMISSIONS } from 'proton-shared/lib/constants';
+
+import Page from '../components/Page';
+
+const { ADMIN, MULTI_USERS } = PERMISSIONS;
+
+export const getMembersPage = () => {
+    return {
+        text: c('Title').t`Users`,
+        route: '/settings/members',
+        permissions: [ADMIN, MULTI_USERS]
+    };
+};
 
 const MembersContainer = () => {
-    return <MembersSection />;
+    return (
+        <Page config={getMembersPage()}>
+            <MembersSection />
+        </Page>
+    );
 };
 
 export default MembersContainer;
