+++ src/app/containers/OrganizationContainer.js
@@ -1,11 +1,35 @@
 import React from 'react';
-import { ObserverSections, OrganizationSection } from 'react-components';
+import { OrganizationSection } from 'react-components';
+import { c } from 'ttag';
+import { PERMISSIONS } from 'proton-shared/lib/constants';
+
+import Page from '../components/Page';
+
+const { ADMIN } = PERMISSIONS;
+
+export const getOrganizationPage = () => {
+    return {
+        text: c('Title').t`Organization`,
+        route: '/settings/organization',
+        permissions: [ADMIN],
+        sections: [
+            {
+                text: c('Title').t`Name`,
+                id: 'name'
+            },
+            {
+                text: c('Title').t`Password & key`,
+                id: 'password'
+            }
+        ]
+    };
+};
 
 const OrganizationContainer = () => {
     return (
-        <ObserverSections>
-            <OrganizationSection id="name" />
-        </ObserverSections>
+        <Page config={getOrganizationPage()}>
+            <OrganizationSection />
+        </Page>
     );
 };
