--- src/app/pages/account.js
@@ -1,34 +1,0 @@
-import { c } from 'ttag';
-
-export const getAccountPage = () => {
-    return {
-        text: c('Title').t`Account`,
-        route: '/settings/account',
-        sections: [
-            {
-                text: c('Title').t`Username`,
-                id: 'username'
-            },
-            {
-                text: c('Title').t`Passwords`,
-                id: 'passwords'
-            },
-            {
-                text: c('Title').t`Two factor authentication`,
-                id: '2fa'
-            },
-            {
-                text: c('Title').t`Recovery & notification`,
-                id: 'email'
-            },
-            {
-                text: c('Title').t`Email subscriptions`,
-                id: 'news'
-            },
-            {
-                text: c('Title').t`Delete account`,
-                id: 'delete'
-            }
-        ]
-    };
-};
