--- src/app/pages/addresses.js
@@ -1,8 +1,0 @@
-import { c } from 'ttag';
-
-export const getAddressesPage = () => {
-    return {
-        text: c('Title').t`Addresses`,
-        route: '/settings/addresses'
-    };
-};
