--- src/app/pages/filters.js
@@ -1,18 +1,0 @@
-import { c } from 'ttag';
-
-export const getFiltersPage = () => {
-    return {
-        text: c('Title').t`Filters`,
-        route: '/settings/filters',
-        sections: [
-            {
-                text: c('Title').t`Custom filters`,
-                id: 'custom'
-            },
-            {
-                text: c('Title').t`Spam filters`,
-                id: 'spam'
-            }
-        ]
-    };
-};
