--- src/app/pages/general.js
@@ -1,34 +1,0 @@
-import { c } from 'ttag';
-
-export const getGeneralPage = () => {
-    return {
-        text: c('Title').t`General`,
-        route: '/settings/general',
-        sections: [
-            {
-                text: c('Title').t`Language`,
-                id: 'language'
-            },
-            {
-                text: c('Title').t`Desktop notifications`,
-                id: 'desktop-notifications'
-            },
-            {
-                text: c('Title').t`Messages`,
-                id: 'messages'
-            },
-            {
-                text: c('Title').t`Contacts`,
-                id: 'contacts'
-            },
-            {
-                text: c('Title').t`Search`,
-                id: 'search'
-            },
-            {
-                text: c('Title').t`Shortcuts`,
-                id: 'shortcuts'
-            }
-        ]
-    };
-};
