--- src/app/pages/helpers.js
@@ -1,8 +1,0 @@
-export const hasPermission = (userPermissions = [], requirements = []) => {
-    if (!requirements.length) {
-        return true;
-    }
-    return requirements.every((permission) => {
-        return userPermissions.includes(permission);
-    });
-};
