--- src/app/pages/labels.js
@@ -1,8 +1,0 @@
-import { c } from 'ttag';
-
-export const getLabelsPage = () => {
-    return {
-        text: c('Title').t`Folders/Labels`,
-        route: '/settings/labels'
-    };
-};
